+++ common/src/main/java/net/conczin/immersive_optimization/TickScheduler.java
@@ -106,7 +106,7 @@ public class TickScheduler {
         public int stressedTicks = 0;
         public int lifeTimeStressedTicks = 0;
 
-        public Int2IntOpenHashMap priorities = new Int2IntOpenHashMap();
+        public volatile Int2IntOpenHashMap priorities = new Int2IntOpenHashMap();
 
         public Map<Long, Integer> blockEntityPriorities = new ConcurrentHashMap<>();
 
@@ -181,7 +181,7 @@ public class TickScheduler {
 
         // Entity culling disabled
         if (!Config.getInstance().enableEntities) {
-            data.priorities.clear();
+            data.priorities = new Int2IntOpenHashMap();
             return;
         }
 
@@ -222,6 +222,7 @@ public class TickScheduler {
         if (!config.entities.getOrDefault(id.toString(), true)) return 0;
         if (!config.entities.getOrDefault(id.getNamespace(), true)) return 0;
         if (!config.cullProjectiles && entity instanceof Projectile) return 0;
+        if (isForceLoaded(level, entity.chunkPosition().toLong())) return 0;
 
         // Find the closest player
         double minDistance = 999999.0;
@@ -277,8 +278,15 @@ public class TickScheduler {
         if (data == null) {
             return true;
         }
+        if (isForceLoaded(level, pos)) {
+            return true;
+        }
         int priority = data.blockEntityPriorities.computeIfAbsent(pos, p -> this.getBlockEntityPriority(level, p));
         return priority < 1 || (level.getGameTime() + pos) % priority == 0;
+    }
+
+    private boolean isForceLoaded(Level level, long chunk) {
+        return !Config.getInstance().optimizeForceLoadedChunks && level instanceof ServerLevel serverLevel && CommonClass.isForceLoaded(serverLevel, chunk);
     }
 
     private int getBlockEntityPriority(Level level, long p) {