+++ changelog.md
@@ -1,3 +1,9 @@
+# 0.1.5
+
+* Fixed rare concurrent access crashes
+* Support force-loaded chunks
+* Disabled block entity optimization by default since it causes too many issues
+
 # 0.1.4
 
 * Added permission levels to commands